refactor: APIDataLoader 구현을 통해 네트워크 통신 기능 분리
